+++ src/holofuel/components/FlashMessage/FlashMessage.module.css
@@ -9,6 +9,7 @@
   align-items: center;
   align-self: center;
   border-radius: 4px;
+  z-index: 30;
 }
 
 .message {
