+++ src/holofuel/contexts/__mocks__/useFlashMessageContext.js
@@ -1,0 +1,9 @@
+const message = 'mock message'
+const time = 123
+export const newMessage = jest.fn()
+
+export default () => ({
+  message,
+  time,
+  newMessage
+})
