--- src/holofuel/contexts/flashMessage.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react'
-
-const FlashMessageContext = createContext()
-
-export default FlashMessageContext
